+++ apx_sdk/src/agent.rs
@@ -10,7 +10,6 @@ pub struct FederationAgent {
     pub response_size_limit: usize,
     pub fetcher_timeout: u64,
     pub deliverer_timeout: u64,
-    pub deliverer_log_response_length: usize,
 
     // Proxy for outgoing requests
     pub proxy_url: Option<String>,